% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\docType{data}
\name{aminoCols}
\alias{aminoCols}
\title{Some colors for amino acids}
\format{A vector with each element giving a color for an amino acid. Vector names correspond to single letter amino acid codes.}
\source{
system.file("data-raw", "makeAminoColors.R", package = "dnaplotr")
}
\usage{
aminoCols
}
\description{
A vector indexed by single letter amino acid code giving color codes based on colors provided by Jmol tweaked slightly so that no two colors are identical
}
\references{
\url{http://jmol.sourceforge.net/jscolors/}
}
\keyword{datasets}

