% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{createFakeDNA}
\alias{createFakeDNA}
\alias{createFakeAA}
\title{Create fake DNA or amino acid sequences}
\usage{
createFakeDNA(n = 500, nChar = 400, nSplit = 3, pGap = 0.3,
  pNoise = 0.01, pMutation = 0.005, bases = c("A", "C", "T", "G", "-"),
  excludeBases = c("-", "X"))

createFakeAA(n = 100, nChar = 100, ..., pGap = 0.2, pNoise = 0,
  pMutation = 0.01, bases = c(names(dnaplotr::aminoCols), "-"))
}
\arguments{
\item{n}{number of fake sequences to generate}

\item{nChar}{character length of the output fake sequences}

\item{nSplit}{The number of hierarchical splits to make in the data e.g. 3 splits produces 2^3=8 "species"}

\item{pGap}{probability of a large insertion or deletion in each grouping}

\item{pNoise}{probability of a random substitution at each base}

\item{pMutation}{probability of a substitution at each base in each hierarchical grouping}

\item{bases}{the bases used in generating the sequence (bases listed in excludeBases are excluded from the initial reference sequence generation}

\item{excludeBases}{bases excluded from the initial reference sequence generation (default: '-' and 'X')}

\item{...}{additional arguments to createFakeDNA}
}
\value{
A n+1 length character vector of fake sequences. The first sequence is the reference. Names of the remaining sequences indicate their hierarchical groupings follow by an arbitrary id
}
\description{
Creates a random reference sequence then adds mutations and indels to
hierarchical sets of the sequences and random noise
}
\examples{
createFakeDNA(10,10)
createFakeAA(10,10)
}
