% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{noGapToGap}
\alias{noGapToGap}
\title{Convenience function to convert ungapped coordinates in a reference function to what the coordinates would be in the gapped reference sequence}
\usage{
noGapToGap(refSeq, coords, gapChars = c("*", ".", "-"))
}
\arguments{
\item{refSeq}{the sequence containing gaps}

\item{coords}{coordinates on the ungapped refSeq to be converted into equivalent gapped coordinates}

\item{gapChars}{characters interpreted as gaps}
}
\value{
A vector of coordinates in the gapped reference sequence
}
\description{
Take a reference sequence with gaps and coordinates in the ungapped reference sequence and translate the coordinates to the corresponding positions in the gapped sequence
}
\examples{
noGapToGap('AA--AA-A',c(1:5))
}

