% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{removeGapCols}
\alias{removeGapCols}
\title{Convenience function to remove columns composed mostly of gaps}
\usage{
removeGapCols(seqs, gapChars = c("*", "-", "."), maxGapProp = 0.9)
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{gapChars}{a vector of single characters that count as gaps}

\item{maxGapProp}{remove columns with greater than this proportion of gaps}
}
\value{
A character vector of sequences with gap columns removed
}
\description{
Remove columns in an alignment that contain more than a give proportion of gaps
(e.g. A-A, G-G to AA,GG).
}
\examples{
removeGapCols(c('A-A-','A-AA','A-AT','A-AG'))
}

