% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{replaceAfterStop}
\alias{replaceAfterStop}
\title{Convenience function to remove amino acids following a stop codon}
\usage{
replaceAfterStop(seqs, stopChars = "X", replaceChar = "X")
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{stopChars}{a vector of single characters that count as stop codons}

\item{replaceChar}{character to replace with (default: X)}
}
\value{
A character vector of sequences with amino acids following a stop replaced
}
\description{
Replace any non X characters following a stop codon (marked as X)
with Xs (e.g. LSYXAAA to LSYXXXX). This can be used to frame shift
mutations and early terminations
}
\examples{
replaceAfterStop(c('LYSXAAA','LYSRAAA','AXAAA','AAAAX'))
}

