% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{replaceOuterGaps}
\alias{replaceOuterGaps}
\title{Convenience function to replace gaps at the start and end of a sequence with a different character}
\usage{
replaceOuterGaps(
  seqs,
  leftEnd = TRUE,
  rightEnd = TRUE,
  gapChars = c("*", "-"),
  replaceChar = "."
)
}
\arguments{
\item{seqs}{a character vector of sequences}

\item{leftEnd}{logical indicating whether gaps should be replaced at the start of the sequence}

\item{rightEnd}{logical indicating whether gaps should be replaced at the end of the sequence}

\item{gapChars}{a vector of single characters that count as gaps}

\item{replaceChar}{character to replace start/end gaps with}
}
\value{
A character vector of sequences with start and/or end gaps replaced with replaceChar
}
\description{
Replace gaps at the start and/or end of a sequence with another character
(e.g. ---A-A--A--- to ...A-A--A...). This can be used to indicate the
difference between indels within known sequences and unknown sequence
surrounding a sequence.
}
\examples{
replaceOuterGaps(c('--A-A--','AAA--AAA','--A-A','A-A--'))
}
