% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{seqSplit}
\alias{seqSplit}
\title{Convenience function for binding a bunch of sequences together}
\usage{
seqSplit(..., fill = NULL)
}
\arguments{
\item{...}{Character vectors to turn into a matrix}

\item{fill}{A character to fill the ends of short strings with. If NULL then all strings must be the same length.}
}
\value{
A matrix with a single character in each cell with rows corresponding to the number of strings and columns corresponding to the position in the string
}
\description{
Take a vector of strings and return a matrix with each row corresponding to
a string and each column a position in those strings. If fill is set then
strings are padded to an equal length. Otherwise errors out if strings are
unequal length.
}
\examples{
seqSplit('AAAT','AA','ATT',fill='-')
}

