% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{createFakeDNA}
\alias{createFakeDNA}
\title{Create fake DNA sequences}
\usage{
createFakeDNA(n = 500, nChar = 400, nSplit = 3, pGap = 0.3,
  pNoise = 0.01, pMutation = 0.005, bases = c("A", "C", "T", "G", "-"))
}
\arguments{
\item{n}{number of fake sequences to generate}

\item{nChar}{character length of the output fake sequences}

\item{nSplit}{The number of hierarchical splits to make in the data e.g. 3 splits produces 2^3=8 "species"}

\item{pGap}{probability of a large insertion or deletion in each grouping}

\item{pNoise}{probability of a random substitution at each base}

\item{pMutation}{probability of a substitution at each base in each hierarchical grouping}

\item{bases}{the bases used in generating the sequence. '-' is excluded from the initial reference sequence generation}
}
\value{
A n+1 length character vector of fake sequences. The first sequence is the reference. Names of the remaining sequences indicate their hierarchical groupings follow by an arbitrary id
}
\description{
Creates a random reference sequence then adds mutations and indels to
hierarchical sets of the sequences and random noise
}
\examples{
createFakeDNA(10,10)
}

