% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDNA.R
\docType{package}
\name{dnaplotr}
\alias{dnaplotr}
\alias{dnaplotr-package}
\title{Functions to plot out a collection of DNA sequences}
\description{
Produce a visual representation of a vector of DNA sequences
}
\details{
The main function is:
     \describe{
       \item{\code{\link{plotDNA}}:}{to produce a plot from a vector of DNA sequences}
     }

And main helper functions are:
     \describe{
       \item{\code{\link{replaceOuterGaps}}:}{to mark gaps at the ends of sequences differently than internal indels}
     }

A vignette is a available using \code{vignette('example',package='dnaplotr')}.
}
\examples{
seqs<-c('ACTGGA','ACTGCA','ACTGGC')
plotDNA(seqs)
}
\author{
Scott Sherrill-Mix, \email{shescott@upenn.edu}
}

