% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\name{plotDNA}
\alias{plotAA}
\alias{plotDNA}
\title{Plot a bunch of DNA sequences}
\usage{
plotDNA(seqs, seqCounts = rep(1, length(seqs)), cols = c(A = "green", T =
  "red", C = "blue", G = "yellow", `-` = "grey", default = "white"),
  xlab = "Position", ylab = "Sequence Read", display = c(groups =
  !is.null(groups)), xStart = 1, groups = NULL, groupCexScale = FALSE,
  refSeq = NULL, ...)

plotAA(..., cols = c(DNAPlotR::aminoCols, `-` = "grey", X = "black"))
}
\arguments{
\item{seqs}{A character vector containing DNA sequences}

\item{seqCounts}{A integer vector with the number of counts for each sequence. This can be used to improve run time and file size if some sequences are duplicated multiple times (default: 1 for each entry in seqs)}

\item{cols}{A named vector with names corresponding to the DNA bases and values showing the appropriate color (default: A: green, T: red, C: blue, G: yellow, -: grey)}

\item{xlab}{A string specifying the x-axis label (default: Position)}

\item{ylab}{A string specifying the y-axis label (default: Sequence read)}

\item{display}{A logical vector with element names in 'legend', 'xAxis', 'yAxis', 'groups' where a FALSE suppresses outputting that plot element (default: TRUE for all or any missing elements)}

\item{xStart}{First base in plot should be labelled as this (default: 1)}

\item{groups}{Group sequences by group and show label on right side of plot. Note that any prior ordering of sequences will be disrupted. Use a factor and reorder the levels to set a particular order of groups.}

\item{groupCexScale}{A logical wheter to scale group label size by the number of sequences. Useful to highlight more abundant groups and help squeeze in labels on smaller groups.}

\item{refSeq}{Reference sequence used for numbering the x-axis without counting gaps present in this sequence}

\item{...}{Additional arguments to plot}
}
\description{
Take a vector of strings representing DNA sequences and plot them to the
current device. A, C, T and G are colored, - are colored gray and all other
characters are white.
}
\section{Functions}{
\itemize{
\item \code{plotAA}: Plot a bunch of AA sequences
}}
\examples{
plotDNA(c('ACACA','ACACA','ACACT'))
refSeq<-'AC---A'
seqs<-c('ACTGGA','ACTGCA','ACTGGC','GCTGGG','GGGG',refSeq)
par(mar=c(4,5,.5,6),cex.axis=2,cex.lab=2)
groupOrder<-c('Group2','Group1','Group3','Reference')
groups<-factor(c('Group1','Group2','Group3','Group1','Group3','Reference'),levels=groupOrder)
seqCounts<-c(30,10,10,15,5,5)
plotDNA(seqs,seqCounts=seqCounts,groups=groups,xStart=10,groupCexScale=TRUE,refSeq=refSeq)
fakeSeqs<-createFakeDNA(1000)
refSeq<-fakeSeqs[1]
fakeSeqs<-fakeSeqs[-1]
species<-sprintf('Species \%s',sub(' [0-9]+$','',names(fakeSeqs)))
par(mar=c(3.5,4.4,.5,7))
plotDNA(fakeSeqs,groups=species,groupCexScale=TRUE)
fakeAA<-c('MALWTRLRPLLALLALWPPPPARAFVNQHLCGSHLVEALY',
'MALWTRLRPLLALLALWPPPPARAFVNX')
plotAA(fakeAA)
}

