% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotDNA.R
\docType{package}
\name{plotDNA}
\alias{plotDNA}
\alias{plotDNA-package}
\title{Functions to plot out a collection of DNA sequences}
\usage{
plotDNA(seqs, emptyTrim = TRUE, gapTrim = 0, groups = NULL,
  groupOrdering = c(), legend = !noText, endGapRemove = FALSE,
  orderBy = NULL, pause = FALSE, extraCmds = NULL, xstart = 1,
  distOrderDecreasing = FALSE, refSeq = NULL, groupCex = NULL,
  lineStagger = FALSE, groupCexScale = FALSE, convertGap2NoGap = FALSE,
  seqCounts = rep(1, length(seqs)), fixedAxis = NULL, refGapWhite = FALSE,
  noText = FALSE, xlab = "Position", ylab = "Sequence Read",
  noTick = FALSE, seqCountDisplay = TRUE, maxAxis = Inf, ...)
}
\arguments{
\item{seqs}{A character vector containing DNA sequences}
}
\value{
A data frame with columns:
     \describe{
       \item{start:}{Start of a contiguous region}
       \item{end:}{End of a contiguous region}
     }
}
\description{
Produce a visual representation of a vector of DNA sequences

Take a vector of strings representing DNA sequences and plot them to the current device. A, C, T and G are colored, - are colored gray and all other characters are white.
}
\details{
The main function is:
     \describe{
       \item{\code{\link{plotDNA}}:}{to produce a plot from a vector of DNA sequences}
     }
}
\examples{
seqs<-c('ACTGGA','ACTGCA','ACTGGC')
plotDNA(seqs)
plotDNA:::index2range(c(1:10,11,15,16,17:20))
}
\author{
Scott Sherrill-Mix, \email{shescott@upenn.edu}
}

